backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.5.1
cryptography==42.0.7
cssselect==1.2.0
defusedxml==0.8.0rc2
distlib==0.3.8
Django==4.2.13
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-filter==24.2
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
django-with-asserts==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
Faker==25.4.0
filelock==3.13.3
idna==3.7
importlib_metadata==7.1.0
lxml==5.2.2
Markdown==3.6
oauthlib==3.2.2
optional-django==0.1.0
platformdirs==4.2.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.2
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
zipp==3.18.2
